Move sequence, iterable, deque, dict/mapping family and None tests to tests/types/
New files: test_dict.py, test_ordered_dict.py, test_defaultdict.py,
test_counter.py, test_mapping.py, test_sequence.py, test_iterable.py,
test_deque.py, test_none.py. Also move the remaining bare
list/set/tuple tests to their existing files.
