+++ tests/test_types.py
@@ -133,13 +133,6 @@ except ImportError:
 # TODO add back tests for Iterator
 
 
-def test_optional():
-    ta = TypeAdapter(Optional[int])  # noqa: UP045
-
-    assert ta.validate_python(None) is None
-    assert ta.validate_python(1) == 1
-
-
 def test_strict_raw_type():
     class Model(BaseModel):
         v: Annotated[str, Strict]
@@ -522,535 +515,6 @@ def test_email_validator_not_installed_name_email():
 
         class Model(BaseModel):
             str_email: NameEmail = ...
-
-
-def test_dict():
-    class Model(BaseModel):
-        v: dict
-
-    assert Model(v={1: 10, 2: 20}).v == {1: 10, 2: 20}
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=[(1, 2), (3, 4)])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'dict_type',
-            'loc': ('v',),
-            'msg': 'Input should be a valid dictionary',
-            'input': [(1, 2), (3, 4)],
-        }
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=[1, 2, 3])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'dict_type', 'loc': ('v',), 'msg': 'Input should be a valid dictionary', 'input': [1, 2, 3]}
-    ]
-
-
-@pytest.mark.parametrize(
-    'value,result',
-    (
-        ([1, 2, '3'], [1, 2, '3']),
-        ((1, 2, '3'), [1, 2, '3']),
-        pytest.param((i**2 for i in range(5)), [0, 1, 4, 9, 16], marks=pytest.mark.thread_unsafe),
-        (deque([1, 2, 3]), [1, 2, 3]),
-        ({1, '2'}, IsOneOf([1, '2'], ['2', 1])),
-    ),
-)
-def test_list_success(value, result):
-    class Model(BaseModel):
-        v: list
-
-    assert Model(v=value).v == result
-
-
-@pytest.mark.parametrize('value', (pytest.param(123, id='int-123'), pytest.param('123', id='str-123')))
-def test_list_fails(value):
-    class Model(BaseModel):
-        v: list
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'list_type',
-            'loc': ('v',),
-            'msg': 'Input should be a valid list',
-            'input': value,
-        }
-    ]
-
-
-def test_ordered_dict():
-    class Model(BaseModel):
-        v: OrderedDict
-
-    assert Model(v=OrderedDict([(1, 10), (2, 20)])).v == OrderedDict([(1, 10), (2, 20)])
-    assert Model(v={1: 10, 2: 20}).v == OrderedDict([(1, 10), (2, 20)])
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=[1, 2, 3])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'dict_type', 'loc': ('v',), 'msg': 'Input should be a valid dictionary', 'input': [1, 2, 3]}
-    ]
-
-
-@pytest.mark.parametrize(
-    'value,result',
-    (
-        ([1, 2, '3'], (1, 2, '3')),
-        ((1, 2, '3'), (1, 2, '3')),
-        pytest.param((i**2 for i in range(5)), (0, 1, 4, 9, 16), marks=pytest.mark.thread_unsafe),
-        (deque([1, 2, 3]), (1, 2, 3)),
-        ({1, '2'}, IsOneOf((1, '2'), ('2', 1))),
-    ),
-)
-def test_tuple_success(value, result):
-    class Model(BaseModel):
-        v: tuple
-
-    assert Model(v=value).v == result
-
-
-@pytest.mark.parametrize('value', (pytest.param(123, id='int-123'), pytest.param('123', id='str-123')))
-def test_tuple_fails(value):
-    class Model(BaseModel):
-        v: tuple
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'tuple_type', 'loc': ('v',), 'msg': 'Input should be a valid tuple', 'input': value}
-    ]
-
-
-@pytest.mark.parametrize(
-    'value,cls,result',
-    (
-        ([1, 2, '3'], int, (1, 2, 3)),
-        ((1, 2, '3'), int, (1, 2, 3)),
-        pytest.param((i**2 for i in range(5)), int, (0, 1, 4, 9, 16), marks=pytest.mark.thread_unsafe),
-        (('a', 'b', 'c'), str, ('a', 'b', 'c')),
-    ),
-)
-def test_tuple_variable_len_success(value, cls, result):
-    class Model(BaseModel):
-        v: tuple[cls, ...]
-
-    assert Model(v=value).v == result
-
-
-@pytest.mark.parametrize(
-    'value, cls, exc',
-    [
-        (
-            ('a', 'b', [1, 2], 'c'),
-            str,
-            [
-                {
-                    'type': 'string_type',
-                    'loc': ('v', 2),
-                    'msg': 'Input should be a valid string',
-                    'input': [1, 2],
-                }
-            ],
-        ),
-        (
-            ('a', 'b', [1, 2], 'c', [3, 4]),
-            str,
-            [
-                {
-                    'type': 'string_type',
-                    'loc': ('v', 2),
-                    'msg': 'Input should be a valid string',
-                    'input': [1, 2],
-                },
-                {
-                    'type': 'string_type',
-                    'loc': ('v', 4),
-                    'msg': 'Input should be a valid string',
-                    'input': [3, 4],
-                },
-            ],
-        ),
-    ],
-)
-def test_tuple_variable_len_fails(value, cls, exc):
-    class Model(BaseModel):
-        v: tuple[cls, ...]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=value)
-    assert exc_info.value.errors(include_url=False) == exc
-
-
-@pytest.mark.parametrize(
-    'value,result',
-    (
-        ({1, 2, '3'}, {1, 2, '3'}),
-        ((1, 2, 2, '3'), {1, 2, '3'}),
-        ([1, 2, 2, '3'], {1, 2, '3'}),
-        ({i**2 for i in range(5)}, {0, 1, 4, 9, 16}),
-    ),
-)
-def test_set_success(value, result):
-    class Model(BaseModel):
-        v: set
-
-    assert Model(v=value).v == result
-
-
-@pytest.mark.parametrize('value', (pytest.param(123, id='int-123'), pytest.param('123', id='str-123')))
-def test_set_fails(value):
-    class Model(BaseModel):
-        v: set
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'set_type', 'loc': ('v',), 'msg': 'Input should be a valid set', 'input': value}
-    ]
-
-
-def test_list_type_fails():
-    class Model(BaseModel):
-        v: list[int]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v='123')
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'list_type', 'loc': ('v',), 'msg': 'Input should be a valid list', 'input': '123'}
-    ]
-
-
-def test_set_type_fails():
-    class Model(BaseModel):
-        v: set[int]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v='123')
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'set_type', 'loc': ('v',), 'msg': 'Input should be a valid set', 'input': '123'}
-    ]
-
-
-@pytest.mark.parametrize(
-    'cls, value,result',
-    (
-        (int, [1, 2, 3], [1, 2, 3]),
-        (int, (1, 2, 3), (1, 2, 3)),
-        (int, range(5), [0, 1, 2, 3, 4]),
-        (int, deque((1, 2, 3)), deque((1, 2, 3))),
-        (set[int], [{1, 2}, {3, 4}, {5, 6}], [{1, 2}, {3, 4}, {5, 6}]),
-        (tuple[int, str], ((1, 'a'), (2, 'b'), (3, 'c')), ((1, 'a'), (2, 'b'), (3, 'c'))),
-    ),
-)
-def test_sequence_success(cls, value, result):
-    class Model(BaseModel):
-        v: Sequence[cls]
-
-    assert Model(v=value).v == result
-
-
-def int_iterable():
-    i = 0
-    while True:
-        i += 1
-        yield str(i)
-
-
-def str_iterable():
-    while True:
-        yield from 'foobarbaz'
-
-
-def test_infinite_iterable_int():
-    class Model(BaseModel):
-        it: Iterable[int]
-
-    m = Model(it=int_iterable())
-
-    assert repr(m.it) == 'ValidatorIterator(index=0, schema=Some(Int(IntValidator { strict: false })))'
-
-    output = []
-    for i in m.it:
-        output.append(i)
-        if i == 10:
-            break
-
-    assert output == [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-
-    m = Model(it=[1, 2, 3])
-    assert list(m.it) == [1, 2, 3]
-
-    m = Model(it=str_iterable())
-    with pytest.raises(ValidationError) as exc_info:
-        next(m.it)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': (0,),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'f',
-        }
-    ]
-
-
-@pytest.mark.parametrize(
-    'type_annotation', (pytest.param(Iterable[Any], id='Iterable[Any]'), pytest.param(Iterable, id='Iterable'))
-)
-def test_iterable_any(type_annotation):
-    class Model(BaseModel):
-        it: type_annotation
-
-    m = Model(it=int_iterable())
-
-    output = []
-    for i in m.it:
-        output.append(i)
-        if int(i) == 10:
-            break
-
-    assert output == ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10']
-
-    m = Model(it=[1, '2', b'three'])
-    assert list(m.it) == [1, '2', b'three']
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(it=3)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'iterable_type', 'loc': ('it',), 'msg': 'Input should be iterable', 'input': 3}
-    ]
-
-    class MaxLenModel(BaseModel):
-        it: type_annotation = Field(max_length=2)
-
-    m = MaxLenModel(it=[1, 2, 3])
-    with pytest.raises(ValidationError) as exc_info:
-        list(m.it)
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_long',
-            'loc': (),
-            'msg': 'Generator should have at most 2 items after validation, not more',
-            'input': [1, 2, 3],
-            'ctx': {'field_type': 'Generator', 'max_length': 2, 'actual_length': None},
-        }
-    ]
-
-
-def test_invalid_iterable():
-    class Model(BaseModel):
-        it: Iterable[int]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(it=3)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'iterable_type', 'loc': ('it',), 'msg': 'Input should be iterable', 'input': 3}
-    ]
-
-
-@pytest.mark.parametrize(
-    'config,input_str',
-    (
-        ({}, 'type=iterable_type, input_value=5, input_type=int'),
-        ({'hide_input_in_errors': False}, 'type=iterable_type, input_value=5, input_type=int'),
-        ({'hide_input_in_errors': True}, 'type=iterable_type'),
-    ),
-)
-def test_iterable_error_hide_input(config, input_str):
-    class Model(BaseModel):
-        it: Iterable[int]
-
-        model_config = ConfigDict(**config)
-
-    with pytest.raises(ValidationError, match=re.escape(f'Input should be iterable [{input_str}]')):
-        Model(it=5)
-
-
-def test_infinite_iterable_validate_first():
-    class Model(BaseModel):
-        it: Iterable[int]
-        b: int
-
-        @field_validator('it')
-        @classmethod
-        def infinite_first_int(cls, it):
-            return itertools.chain([next(it)], it)
-
-    m = Model(it=int_iterable(), b=3)
-
-    assert m.b == 3
-    assert m.it
-
-    for i in m.it:
-        assert i
-        if i == 10:
-            break
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(it=str_iterable(), b=3)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('it', 0),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'f',
-        }
-    ]
-
-
-def test_sequence_generator_fails():
-    class Model(BaseModel):
-        v: Sequence[int]
-
-    gen = (i for i in [1, 2, 3])
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=gen)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'is_instance_of',
-            'loc': ('v',),
-            'msg': 'Input should be an instance of Sequence',
-            'input': gen,
-            'ctx': {'class': 'Sequence'},
-        }
-    ]
-
-
-@pytest.mark.parametrize(
-    'cls,value,errors',
-    (
-        (
-            int,
-            [1, 'a', 3],
-            [
-                {
-                    'type': 'int_parsing',
-                    'loc': ('v', 1),
-                    'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                    'input': 'a',
-                },
-            ],
-        ),
-        (
-            int,
-            (1, 2, 'a'),
-            [
-                {
-                    'type': 'int_parsing',
-                    'loc': ('v', 2),
-                    'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                    'input': 'a',
-                },
-            ],
-        ),
-        (
-            float,
-            ('a', 2.2, 3.3),
-            [
-                {
-                    'type': 'float_parsing',
-                    'loc': ('v', 0),
-                    'msg': 'Input should be a valid number, unable to parse string as a number',
-                    'input': 'a',
-                },
-            ],
-        ),
-        (
-            float,
-            (1.1, 2.2, 'a'),
-            [
-                {
-                    'type': 'float_parsing',
-                    'loc': ('v', 2),
-                    'msg': 'Input should be a valid number, unable to parse string as a number',
-                    'input': 'a',
-                },
-            ],
-        ),
-        (
-            float,
-            {1.0, 2.0, 3.0},
-            [
-                {
-                    'type': 'is_instance_of',
-                    'loc': ('v',),
-                    'msg': 'Input should be an instance of Sequence',
-                    'input': {
-                        1.0,
-                        2.0,
-                        3.0,
-                    },
-                    'ctx': {
-                        'class': 'Sequence',
-                    },
-                },
-            ],
-        ),
-        (
-            set[int],
-            [{1, 2}, {2, 3}, {'d'}],
-            [
-                {
-                    'type': 'int_parsing',
-                    'loc': ('v', 2, 0),
-                    'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                    'input': 'd',
-                }
-            ],
-        ),
-        (
-            tuple[int, str],
-            ((1, 'a'), ('a', 'a'), (3, 'c')),
-            [
-                {
-                    'type': 'int_parsing',
-                    'loc': ('v', 1, 0),
-                    'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                    'input': 'a',
-                }
-            ],
-        ),
-        (
-            list[int],
-            [{'a': 1, 'b': 2}, [1, 2], [2, 3]],
-            [
-                {
-                    'type': 'list_type',
-                    'loc': ('v', 0),
-                    'msg': 'Input should be a valid list',
-                    'input': {'a': 1, 'b': 2},
-                }
-            ],
-        ),
-    ),
-    ids=repr,
-)
-def test_sequence_fails(cls, value, errors):
-    class Model(BaseModel):
-        v: Sequence[cls]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=value)
-    assert exc_info.value.errors(include_url=False) == errors
-
-
-def test_sequence_strict():
-    assert TypeAdapter(Sequence[int]).validate_python((), strict=True) == ()
 
 
 def test_new_type_success():
@@ -2101,348 +1565,6 @@ def test_bytesize_raises():
         m.size.to('1ZiB')
 
 
-def test_deque_success():
-    class Model(BaseModel):
-        v: deque
-
-    assert Model(v=[1, 2, 3]).v == deque([1, 2, 3])
-
-
-@pytest.mark.parametrize(
-    'cls,value,result',
-    (
-        (int, [1, 2, 3], deque([1, 2, 3])),
-        (int, (1, 2, 3), deque((1, 2, 3))),
-        (int, deque((1, 2, 3)), deque((1, 2, 3))),
-        (float, [1.0, 2.0, 3.0], deque([1.0, 2.0, 3.0])),
-        (set[int], [{1, 2}, {3, 4}, {5, 6}], deque([{1, 2}, {3, 4}, {5, 6}])),
-        (tuple[int, str], ((1, 'a'), (2, 'b'), (3, 'c')), deque(((1, 'a'), (2, 'b'), (3, 'c')))),
-        (str, 'one two three'.split(), deque(['one', 'two', 'three'])),
-        (
-            int,
-            {1: 10, 2: 20, 3: 30}.keys(),
-            deque([1, 2, 3]),
-        ),
-        (
-            int,
-            {1: 10, 2: 20, 3: 30}.values(),
-            deque([10, 20, 30]),
-        ),
-        (
-            tuple[int, int],
-            {1: 10, 2: 20, 3: 30}.items(),
-            deque([(1, 10), (2, 20), (3, 30)]),
-        ),
-        (
-            float,
-            {1, 2, 3},
-            deque([1, 2, 3]),
-        ),
-        (
-            float,
-            frozenset((1, 2, 3)),
-            deque([1, 2, 3]),
-        ),
-    ),
-)
-def test_deque_generic_success(cls, value, result):
-    class Model(BaseModel):
-        v: deque[cls]
-
-    assert Model(v=value).v == result
-
-
-@pytest.mark.parametrize(
-    'cls,value,result',
-    (
-        (int, deque((1, 2, 3)), deque((1, 2, 3))),
-        (str, deque(('1', '2', '3')), deque(('1', '2', '3'))),
-    ),
-)
-def test_deque_generic_success_strict(cls, value: Any, result):
-    class Model(BaseModel):
-        v: deque[cls]
-
-        model_config = ConfigDict(strict=True)
-
-    assert Model(v=value).v == result
-
-
-@pytest.mark.parametrize(
-    'cls,value,expected_error',
-    (
-        (
-            int,
-            [1, 'a', 3],
-            {
-                'type': 'int_parsing',
-                'loc': ('v', 1),
-                'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                'input': 'a',
-            },
-        ),
-        (
-            int,
-            (1, 2, 'a'),
-            {
-                'type': 'int_parsing',
-                'loc': ('v', 2),
-                'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                'input': 'a',
-            },
-        ),
-        (
-            tuple[int, str],
-            ((1, 'a'), ('a', 'a'), (3, 'c')),
-            {
-                'type': 'int_parsing',
-                'loc': ('v', 1, 0),
-                'msg': 'Input should be a valid integer, unable to parse string as an integer',
-                'input': 'a',
-            },
-        ),
-        (
-            list[int],
-            [{'a': 1, 'b': 2}, [1, 2], [2, 3]],
-            {
-                'type': 'list_type',
-                'loc': ('v', 0),
-                'msg': 'Input should be a valid list',
-                'input': {
-                    'a': 1,
-                    'b': 2,
-                },
-            },
-        ),
-    ),
-)
-def test_deque_fails(cls, value, expected_error):
-    class Model(BaseModel):
-        v: deque[cls]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=value)
-    # debug(exc_info.value.errors(include_url=False))
-    assert len(exc_info.value.errors(include_url=False)) == 1
-    assert expected_error == exc_info.value.errors(include_url=False)[0]
-
-
-def test_deque_model():
-    class Model2(BaseModel):
-        x: int
-
-    class Model(BaseModel):
-        v: deque[Model2]
-
-    seq = [Model2(x=1), Model2(x=2)]
-    assert Model(v=seq).v == deque(seq)
-
-
-def test_deque_json():
-    class Model(BaseModel):
-        v: deque[int]
-
-    assert Model(v=deque((1, 2, 3))).model_dump_json() == '{"v":[1,2,3]}'
-
-
-def test_deque_any_maxlen():
-    class DequeModel1(BaseModel):
-        field: deque
-
-    assert DequeModel1(field=deque()).field.maxlen is None
-    assert DequeModel1(field=deque(maxlen=8)).field.maxlen == 8
-
-    class DequeModel2(BaseModel):
-        field: deque = deque()
-
-    assert DequeModel2().field.maxlen is None
-    assert DequeModel2(field=deque()).field.maxlen is None
-    assert DequeModel2(field=deque(maxlen=8)).field.maxlen == 8
-
-    class DequeModel3(BaseModel):
-        field: deque = deque(maxlen=5)
-
-    assert DequeModel3().field.maxlen == 5
-    assert DequeModel3(field=deque()).field.maxlen is None
-    assert DequeModel3(field=deque(maxlen=8)).field.maxlen == 8
-
-
-def test_deque_typed_maxlen():
-    class DequeModel1(BaseModel):
-        field: deque[int]
-
-    assert DequeModel1(field=deque()).field.maxlen is None
-    assert DequeModel1(field=deque(maxlen=8)).field.maxlen == 8
-
-    class DequeModel2(BaseModel):
-        field: deque[int] = deque()
-
-    assert DequeModel2().field.maxlen is None
-    assert DequeModel2(field=deque()).field.maxlen is None
-    assert DequeModel2(field=deque(maxlen=8)).field.maxlen == 8
-
-    class DequeModel3(BaseModel):
-        field: deque[int] = deque(maxlen=5)
-
-    assert DequeModel3().field.maxlen == 5
-    assert DequeModel3(field=deque()).field.maxlen is None
-    assert DequeModel3(field=deque(maxlen=8)).field.maxlen == 8
-
-
-def test_deque_enforces_maxlen():
-    class DequeModel1(BaseModel):
-        field: Annotated[deque[int], Field(max_length=3)]
-
-    with pytest.raises(ValidationError):
-        DequeModel1(field=deque([1, 2, 3, 4]))
-
-
-@pytest.mark.parametrize('value_type', (None, type(None)))
-def test_none(value_type):
-    class Model(BaseModel):
-        my_none: value_type
-        my_none_list: list[value_type]
-        my_none_dict: dict[str, value_type]
-        my_json_none: Json[value_type]
-
-    Model(
-        my_none=None,
-        my_none_list=[None] * 3,
-        my_none_dict={'a': None, 'b': None},
-        my_json_none='null',
-    )
-
-    assert Model.model_json_schema() == {
-        'title': 'Model',
-        'type': 'object',
-        'properties': {
-            'my_none': {'type': 'null', 'title': 'My None'},
-            'my_none_list': {'type': 'array', 'items': {'type': 'null'}, 'title': 'My None List'},
-            'my_none_dict': {'type': 'object', 'additionalProperties': {'type': 'null'}, 'title': 'My None Dict'},
-            'my_json_none': {
-                'contentMediaType': 'application/json',
-                'contentSchema': {'type': 'null'},
-                'title': 'My Json None',
-                'type': 'string',
-            },
-        },
-        'required': ['my_none', 'my_none_list', 'my_none_dict', 'my_json_none'],
-    }
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(
-            my_none='qwe',
-            my_none_list=[1, None, 'qwe'],
-            my_none_dict={'a': 1, 'b': None},
-            my_json_none='"a"',
-        )
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'none_required', 'loc': ('my_none',), 'msg': 'Input should be None', 'input': 'qwe'},
-        {'type': 'none_required', 'loc': ('my_none_list', 0), 'msg': 'Input should be None', 'input': 1},
-        {
-            'type': 'none_required',
-            'loc': ('my_none_list', 2),
-            'msg': 'Input should be None',
-            'input': 'qwe',
-        },
-        {
-            'type': 'none_required',
-            'loc': ('my_none_dict', 'a'),
-            'msg': 'Input should be None',
-            'input': 1,
-        },
-        {'type': 'none_required', 'loc': ('my_json_none',), 'msg': 'Input should be None', 'input': 'a'},
-    ]
-
-
-def test_none_literal():
-    class Model(BaseModel):
-        my_none: Literal[None]
-        my_none_list: list[Literal[None]]
-        my_none_dict: dict[str, Literal[None]]
-        my_json_none: Json[Literal[None]]
-
-    Model(
-        my_none=None,
-        my_none_list=[None] * 3,
-        my_none_dict={'a': None, 'b': None},
-        my_json_none='null',
-    )
-
-    # insert_assert(Model.model_json_schema())
-    assert Model.model_json_schema() == {
-        'title': 'Model',
-        'type': 'object',
-        'properties': {
-            'my_none': {'const': None, 'title': 'My None', 'type': 'null'},
-            'my_none_list': {
-                'items': {'const': None, 'type': 'null'},
-                'title': 'My None List',
-                'type': 'array',
-            },
-            'my_none_dict': {
-                'additionalProperties': {'const': None, 'type': 'null'},
-                'title': 'My None Dict',
-                'type': 'object',
-            },
-            'my_json_none': {
-                'contentMediaType': 'application/json',
-                'contentSchema': {'const': None, 'type': 'null'},
-                'title': 'My Json None',
-                'type': 'string',
-            },
-        },
-        'required': ['my_none', 'my_none_list', 'my_none_dict', 'my_json_none'],
-    }
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(
-            my_none='qwe',
-            my_none_list=[1, None, 'qwe'],
-            my_none_dict={'a': 1, 'b': None},
-            my_json_none='"a"',
-        )
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'literal_error',
-            'loc': ('my_none',),
-            'msg': 'Input should be None',
-            'input': 'qwe',
-            'ctx': {'expected': 'None'},
-        },
-        {
-            'type': 'literal_error',
-            'loc': ('my_none_list', 0),
-            'msg': 'Input should be None',
-            'input': 1,
-            'ctx': {'expected': 'None'},
-        },
-        {
-            'type': 'literal_error',
-            'loc': ('my_none_list', 2),
-            'msg': 'Input should be None',
-            'input': 'qwe',
-            'ctx': {'expected': 'None'},
-        },
-        {
-            'type': 'literal_error',
-            'loc': ('my_none_dict', 'a'),
-            'msg': 'Input should be None',
-            'input': 1,
-            'ctx': {'expected': 'None'},
-        },
-        {
-            'type': 'literal_error',
-            'loc': ('my_json_none',),
-            'msg': 'Input should be None',
-            'input': 'a',
-            'ctx': {'expected': 'None'},
-        },
-    ]
-
-
 def test_default_union_types():
     class DefaultModel(BaseModel):
         v: int | bool | str
@@ -2802,239 +1924,6 @@ def test_base64url_invalid(field_type, input_data):
     ]
 
 
-def test_sequence_subclass_without_core_schema() -> None:
-    class MyList(list[int]):
-        # The point of this is that subclasses can do arbitrary things
-        # This is the reason why we don't try to handle them automatically
-        # TBD if we introspect `__init__` / `__new__`
-        # (which is the main thing that would mess us up if modified in a subclass)
-        # and automatically handle cases where the subclass doesn't override it.
-        # There's still edge cases (again, arbitrary behavior...)
-        # and it's harder to explain, but could lead to a better user experience in some cases
-        # It will depend on how the complaints (which have and will happen in both directions)
-        # balance out
-        def __init__(self, *args: Any, required: int, **kwargs: Any) -> None:
-            super().__init__(*args, **kwargs)
-
-    with pytest.raises(
-        PydanticSchemaGenerationError, match='implement `__get_pydantic_core_schema__` on your type to fully support it'
-    ):
-
-        class _(BaseModel):
-            x: MyList
-
-
-def test_typing_coercion_defaultdict():
-    class Model(BaseModel):
-        x: defaultdict[int, str]
-
-    d = defaultdict(str)
-    d['1']
-    m = Model(x=d)
-    assert isinstance(m.x, defaultdict)
-    assert repr(m.x) == "defaultdict(<class 'str'>, {1: ''})"
-
-
-def test_typing_coercion_counter():
-    class Model(BaseModel):
-        x: Counter[str]
-
-    m = Model(x={'a': 10})
-    assert isinstance(m.x, Counter)
-    assert repr(m.x) == "Counter({'a': 10})"
-
-
-def test_typing_counter_value_validation():
-    class Model(BaseModel):
-        x: Counter[str]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(x={'a': 'a'})
-
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('x', 'a'),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'a',
-        }
-    ]
-
-
-def test_mapping_subclass_without_core_schema() -> None:
-    class MyDict(dict[int, int]):
-        # The point of this is that subclasses can do arbitrary things
-        # This is the reason why we don't try to handle them automatically
-        # TBD if we introspect `__init__` / `__new__`
-        # (which is the main thing that would mess us up if modified in a subclass)
-        # and automatically handle cases where the subclass doesn't override it.
-        # There's still edge cases (again, arbitrary behavior...)
-        # and it's harder to explain, but could lead to a better user experience in some cases
-        # It will depend on how the complaints (which have and will happen in both directions)
-        # balance out
-        def __init__(self, *args: Any, required: int, **kwargs: Any) -> None:
-            super().__init__(*args, **kwargs)
-
-    with pytest.raises(
-        PydanticSchemaGenerationError, match='implement `__get_pydantic_core_schema__` on your type to fully support it'
-    ):
-
-        class _(BaseModel):
-            x: MyDict
-
-
-def test_defaultdict_unknown_default_factory() -> None:
-    """
-    https://github.com/pydantic/pydantic/issues/4687
-    """
-    with pytest.raises(
-        PydanticSchemaGenerationError,
-        match=r'Unable to infer a default factory for keys of type collections.defaultdict\[int, int\]',
-    ):
-
-        class Model(BaseModel):
-            d: defaultdict[int, defaultdict[int, int]]
-
-
-def test_defaultdict_infer_default_factory() -> None:
-    class Model(BaseModel):
-        a: defaultdict[int, list[int]]
-        b: defaultdict[int, int]
-        c: defaultdict[int, set]
-
-    m = Model(a={}, b={}, c={})
-    assert m.a.default_factory is not None
-    assert m.a.default_factory() == []
-    assert m.b.default_factory is not None
-    assert m.b.default_factory() == 0
-    assert m.c.default_factory is not None
-    assert m.c.default_factory() == set()
-
-
-def test_defaultdict_explicit_default_factory() -> None:
-    class MyList(list[int]):
-        pass
-
-    class Model(BaseModel):
-        a: defaultdict[int, Annotated[list[int], Field(default_factory=lambda: MyList())]]
-
-    m = Model(a={})
-    assert m.a.default_factory is not None
-    assert isinstance(m.a.default_factory(), MyList)
-
-
-def test_defaultdict_default_factory_preserved() -> None:
-    class Model(BaseModel):
-        a: defaultdict[int, list[int]]
-
-    class MyList(list[int]):
-        pass
-
-    m = Model(a=defaultdict(lambda: MyList()))
-    assert m.a.default_factory is not None
-    assert isinstance(m.a.default_factory(), MyList)
-
-
-def test_custom_default_dict() -> None:
-    KT = TypeVar('KT')
-    VT = TypeVar('VT')
-
-    class CustomDefaultDict(defaultdict[KT, VT]):
-        @classmethod
-        def __get_pydantic_core_schema__(cls, source_type: Any, handler: GetCoreSchemaHandler) -> CoreSchema:
-            keys_type, values_type = get_args(source_type)
-            return core_schema.no_info_after_validator_function(
-                lambda x: cls(x.default_factory, x), handler(defaultdict[keys_type, values_type])
-            )
-
-    ta = TypeAdapter(CustomDefaultDict[str, int])
-
-    assert ta.validate_python({'a': 1}) == CustomDefaultDict(int, {'a': 1})
-
-
-@pytest.mark.parametrize(
-    'field_type',
-    [
-        pytest.param(typing.OrderedDict, id='typing.OrderedDict'),
-        pytest.param(collections.OrderedDict, id='collections.OrderedDict'),
-    ],
-)
-def test_ordered_dict_from_ordered_dict(field_type):
-    class Model(BaseModel):
-        od_field: field_type
-
-    od_value = collections.OrderedDict([('a', 1), ('b', 2)])
-
-    m = Model(od_field=od_value)
-
-    assert isinstance(m.od_field, collections.OrderedDict)
-    assert m.od_field == od_value
-    # we don't make any promises about preserving instances
-    # at the moment we always copy them for consistency and predictability
-    # so this is more so documenting the current behavior than a promise
-    # we make to users
-    assert m.od_field is not od_value
-
-    assert m.model_json_schema() == {
-        'properties': {'od_field': {'title': 'Od Field', 'type': 'object', 'additionalProperties': True}},
-        'required': ['od_field'],
-        'title': 'Model',
-        'type': 'object',
-    }
-
-
-def test_ordered_dict_from_ordered_dict_typed():
-    class Model(BaseModel):
-        od_field: typing.OrderedDict[str, int]
-
-    od_value = collections.OrderedDict([('a', 1), ('b', 2)])
-
-    m = Model(od_field=od_value)
-
-    assert isinstance(m.od_field, collections.OrderedDict)
-    assert m.od_field == od_value
-
-    assert m.model_json_schema() == {
-        'properties': {
-            'od_field': {
-                'additionalProperties': {'type': 'integer'},
-                'title': 'Od Field',
-                'type': 'object',
-            }
-        },
-        'required': ['od_field'],
-        'title': 'Model',
-        'type': 'object',
-    }
-
-
-@pytest.mark.parametrize(
-    'field_type',
-    [
-        pytest.param(typing.OrderedDict, id='typing.OrderedDict'),
-        pytest.param(collections.OrderedDict, id='collections.OrderedDict'),
-    ],
-)
-def test_ordered_dict_from_dict(field_type):
-    class Model(BaseModel):
-        od_field: field_type
-
-    od_value = {'a': 1, 'b': 2}
-
-    m = Model(od_field=od_value)
-
-    assert isinstance(m.od_field, collections.OrderedDict)
-    assert m.od_field == collections.OrderedDict(od_value)
-
-    assert m.model_json_schema() == {
-        'properties': {'od_field': {'title': 'Od Field', 'type': 'object', 'additionalProperties': True}},
-        'required': ['od_field'],
-        'title': 'Model',
-        'type': 'object',
-    }
-
-
 def test_handle_3rd_party_custom_type_reusing_known_metadata() -> None:
     class PdDecimal(Decimal):
         def ___repr__(self) -> str:
@@ -3300,26 +2189,6 @@ def test_transform_schema_for_third_party_class():
             'ctx': {'gt': 123},
         }
     ]
-
-
-def test_iterable_arbitrary_type():
-    class CustomIterable(Iterable):
-        def __init__(self, iterable):
-            self.iterable = iterable
-
-        def __iter__(self):
-            return self
-
-        def __next__(self):
-            return next(self.iterable)
-
-    with pytest.raises(
-        PydanticSchemaGenerationError,
-        match='Unable to generate pydantic-core schema for .*CustomIterable.*. Set `arbitrary_types_allowed=True`',
-    ):
-
-        class Model(BaseModel):
-            x: CustomIterable
 
 
 @pytest.mark.parametrize(
@@ -3803,15 +2672,6 @@ def test_on_error_omit_top_level() -> None:
         ta.validate_python('a')
 
 
-def test_can_serialize_deque_passed_to_sequence() -> None:
-    ta = TypeAdapter(Sequence[int])
-    my_dec = ta.validate_python(deque([1, 2, 3]))
-    assert my_dec == deque([1, 2, 3])
-
-    assert ta.dump_python(my_dec) == my_dec
-    assert ta.dump_json(my_dec) == b'[1,2,3]'
-
-
 @pytest.mark.skipif(
     platform.python_implementation() == 'PyPy',
     reason='PyPy has a bug in complex string parsing. A fix is implemented but not yet released.',
@@ -3967,24 +2827,6 @@ def test_fail_fast(tp, fail_fast, decl) -> None:
     assert exc_info.value.errors(include_url=False) == errors
 
 
-def test_mutable_mapping_userdict_subclass() -> None:
-    """Addresses https://github.com/pydantic/pydantic/issues/9549.
-
-    Note - we still don't do a good job of handling subclasses, as we convert the input to a dict.
-    """
-    adapter = TypeAdapter(MutableMapping, config=ConfigDict(strict=True))
-
-    assert isinstance(adapter.validate_python(UserDict()), MutableMapping)
-
-
-def test_mapping_parameterized() -> None:
-    """https://github.com/pydantic/pydantic/issues/11650"""
-    adapter = TypeAdapter(Mapping[str, int])
-
-    with pytest.raises(ValidationError):
-        adapter.validate_python({'valid': 1, 'invalid': {}})
-
-
 def test_ser_ip_with_union() -> None:
     bool_first = TypeAdapter(Union[bool, ipaddress.IPv4Address])  # noqa: UP007
     assert bool_first.dump_python(True, mode='json') is True
@@ -4088,18 +2930,6 @@ def test_custom_serializer_override_secret_str() -> None:
 
     u = User(name='sam', password='hi')
     assert u.model_dump()['password'] == 'secret: **********'
-
-
-@pytest.mark.parametrize('sequence_type', [list, tuple, deque])
-def test_sequence_with_nested_type(sequence_type: type) -> None:
-    class Model(BaseModel):
-        a: int
-
-    class OuterModel(BaseModel):
-        inner: Sequence[Model]
-
-    models = sequence_type([Model(a=1), Model(a=2)])
-    assert OuterModel(inner=models).model_dump() == {'inner': sequence_type([{'a': 1}, {'a': 2}])}
 
 
 def test_union_respects_local_strict() -> None:
