+++ tests/types/test_dict.py
@@ -1,0 +1,28 @@
+import pytest
+
+from pydantic import BaseModel, ValidationError
+
+
+def test_dict():
+    class Model(BaseModel):
+        v: dict
+
+    assert Model(v={1: 10, 2: 20}).v == {1: 10, 2: 20}
+    with pytest.raises(ValidationError) as exc_info:
+        Model(v=[(1, 2), (3, 4)])
+    # insert_assert(exc_info.value.errors(include_url=False))
+    assert exc_info.value.errors(include_url=False) == [
+        {
+            'type': 'dict_type',
+            'loc': ('v',),
+            'msg': 'Input should be a valid dictionary',
+            'input': [(1, 2), (3, 4)],
+        }
+    ]
+
+    with pytest.raises(ValidationError) as exc_info:
+        Model(v=[1, 2, 3])
+    # insert_assert(exc_info.value.errors(include_url=False))
+    assert exc_info.value.errors(include_url=False) == [
+        {'type': 'dict_type', 'loc': ('v',), 'msg': 'Input should be a valid dictionary', 'input': [1, 2, 3]}
+    ]
