+++ tests/types/test_set.py
@@ -8,6 +8,47 @@ from pydantic import BaseModel, ConfigDict, Field, TypeAdapter, ValidationError
 
 
 @pytest.mark.parametrize(
+    'value,result',
+    (
+        ({1, 2, '3'}, {1, 2, '3'}),
+        ((1, 2, 2, '3'), {1, 2, '3'}),
+        ([1, 2, 2, '3'], {1, 2, '3'}),
+        ({i**2 for i in range(5)}, {0, 1, 4, 9, 16}),
+    ),
+)
+def test_set_success(value, result):
+    class Model(BaseModel):
+        v: set
+
+    assert Model(v=value).v == result
+
+
+@pytest.mark.parametrize('value', (pytest.param(123, id='int-123'), pytest.param('123', id='str-123')))
+def test_set_fails(value):
+    class Model(BaseModel):
+        v: set
+
+    with pytest.raises(ValidationError) as exc_info:
+        Model(v=value)
+    # insert_assert(exc_info.value.errors(include_url=False))
+    assert exc_info.value.errors(include_url=False) == [
+        {'type': 'set_type', 'loc': ('v',), 'msg': 'Input should be a valid set', 'input': value}
+    ]
+
+
+def test_set_type_fails():
+    class Model(BaseModel):
+        v: set[int]
+
+    with pytest.raises(ValidationError) as exc_info:
+        Model(v='123')
+    # insert_assert(exc_info.value.errors(include_url=False))
+    assert exc_info.value.errors(include_url=False) == [
+        {'type': 'set_type', 'loc': ('v',), 'msg': 'Input should be a valid set', 'input': '123'}
+    ]
+
+
+@pytest.mark.parametrize(
     'value,expected',
     [
         ({'1', '2'}, {'1', '2'}),
